Fix brim overlapping with internal holes
If external_only is enabled, then the external brim of a part contained within another part could overlap with the outer part.
This was caused by the fact that polylines are intersected with the allowed_areas,
which meant that the area covered by those polylines could overlap with the disallowed areas by half their line width.